+++ pkg/mimo/model.go
@@ -188,7 +188,7 @@ func (m Metrics) CoherenceRateValidate() int {
 	result := 0
 
 	for _, constraint := range m.Constraints {
-		if constraint.Target == CohenrentRate {
+		if constraint.Target == CoherentRate {
 			if !validate(constraint.Type, constraint.Value, m.Coherence.Rate()) {
 				return -1
 			}
@@ -245,6 +245,72 @@ func (m Metrics) Validate() int {
 	}
 
 	return 0
+}
+
+// GetInvalidSamplesForCoherentRate will return at most n invalid sample if a constraint on coherent rate failed.
+func (m Metrics) GetInvalidSamplesForCoherentRate(maxlen int) []Sample {
+	constraint := m.findFailedCoherentConstraint()
+	samples := []Sample{}
+
+	if constraint != nil {
+		if (constraint.Type == ShouldEqual && constraint.Value > m.Coherence.Rate()) ||
+			constraint.Type == ShouldBeGreaterThan || constraint.Type == ShouldBeGreaterThanOrEqualTo {
+			samples = append(samples, m.Coherence.Backend.GetSamplesMulti(maxlen)...)
+		}
+
+		if (constraint.Type == ShouldEqual && constraint.Value < m.Coherence.Rate()) ||
+			constraint.Type == ShouldBeLessThanOrEqualTo || constraint.Type == ShouldBeLowerThan {
+			samples = append(samples, m.Coherence.Backend.GetSamplesMono(maxlen)...)
+		}
+	}
+
+	return samples
+}
+
+func (m Metrics) findFailedCoherentConstraint() *Constraint {
+	for _, c := range m.Constraints {
+		c := c
+		if c.Target == CoherentRate {
+			if !validate(c.Type, c.Value, m.Coherence.Rate()) {
+				return &c
+			}
+		}
+	}
+
+	return nil
+}
+
+// GetInvalidSamplesForIdentifiantRate will return at most n invalid sample if a constraint on identifiant rate failed.
+func (m Metrics) GetInvalidSamplesForIdentifiantRate(maxlen int) []Sample {
+	constraint := m.findFailedIdentifiantConstraint()
+	samples := []Sample{}
+
+	if constraint != nil {
+		if (constraint.Type == ShouldEqual && constraint.Value > m.Identifiant.Rate()) ||
+			constraint.Type == ShouldBeGreaterThan || constraint.Type == ShouldBeGreaterThanOrEqualTo {
+			samples = append(samples, m.Identifiant.Backend.GetSamplesMulti(maxlen)...)
+		}
+
+		if (constraint.Type == ShouldEqual && constraint.Value < m.Identifiant.Rate()) ||
+			constraint.Type == ShouldBeLessThanOrEqualTo || constraint.Type == ShouldBeLowerThan {
+			samples = append(samples, m.Identifiant.Backend.GetSamplesMono(maxlen)...)
+		}
+	}
+
+	return samples
+}
+
+func (m Metrics) findFailedIdentifiantConstraint() *Constraint {
+	for _, c := range m.Constraints {
+		c := c
+		if c.Target == IdentifiantRate {
+			if !validate(c.Type, c.Value, m.Identifiant.Rate()) {
+				return &c
+			}
+		}
+	}
+
+	return nil
 }
 
 type Report struct {
@@ -450,4 +516,9 @@ func isExcluded(exclude []any, value any, valueStr string) bool {
 	}
 
 	return false
+}
+
+type Sample struct {
+	OriginalValue  string
+	AssignedValues []string
 }